개인 회원 탈퇴 / 어드민 회원 강제탈퇴부분 수정 (게시물/댓글/좋아요 없어야함)

+++ PuppyRun/src/user/controller/DeleteServlet.java
@@ -72,7 +72,7 @@ public class DeleteServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		// 어드민 회원 강제 탈퇴
 		String userId = request.getParameter("userId");
-		int noticeNo = Integer.parseInt(request.getParameter("noticeNo"));
+		String noticeNo = request.getParameter("noticeNo");
 		
 		int userResult = 0;
 		int dogResult = 0;
@@ -81,8 +81,9 @@ public class DeleteServlet extends HttpServlet {
 		int likeResult = 0;
 		
 		System.out.println(noticeNo);
-		if(noticeNo != 0) {
-			int result = new NoticeService().deleteNotice(noticeNo);
+		if(noticeNo != null) {
+			int noticeNo1 = Integer.parseInt(noticeNo);
+			int result = new NoticeService().deleteNotice(noticeNo1);
 			
 			if(result > 0) {
 				response.sendRedirect("/WEB-INF/views/user/myinfo-m.jsp");
@@ -94,30 +95,34 @@ public class DeleteServlet extends HttpServlet {
 		
 		System.out.println("userId 받아왔니 : " + userId);
 		User user = new UserService().selectOneUserIdOnly(userId);
-		if(userId != null) {
+		if(noticeNo == null && userId != null) {
 			if(user.getDogCheck() == 'N') {
 				// 회원 삭제할 때, 게시물과 댓글, 좋아요도 삭제
-				likeResult = new LikeService().deleteLike(userId); 
-				commentResult = new CommentService().deleteCommunity(userId);
-				communityResult = new CommunityService().deleteCommunity(userId);
+				/*
+				 * likeResult = new LikeService().deleteLike(userId); commentResult = new
+				 * CommentService().deleteCommunity(userId); communityResult = new
+				 * CommunityService().deleteCommunity(userId);
+				 */
 				userResult = new UserService().deleteUser(userId);
 			
 				if(userResult > 0 ) {
-					response.sendRedirect("/WEB-INF/views/user/myinfo-m.jsp");
+					response.sendRedirect("/user/list?dogCheck=all");
 				}else {
 					System.out.println("관리자 메뉴에서 회원 삭제 오류 발생 (DogCheck == N)");
 				}
 			// 아래 부분 오류 있음 (자식 레코드)
 			}else if(user.getDogCheck() == 'Y'){
-				likeResult = new LikeService().deleteLike(userId);
-				commentResult = new CommentService().deleteCommunity(userId);
-				communityResult = new CommunityService().deleteCommunity(userId);
+				/*
+				 * likeResult = new LikeService().deleteLike(userId); commentResult = new
+				 * CommentService().deleteCommunity(userId); communityResult = new
+				 * CommunityService().deleteCommunity(userId);
+				 */
 				dogResult = new UserService().deleteDog(userId);
 				userResult = new UserService().deleteUser(userId);
 			
 			
 				if(userResult > 0 && dogResult > 0) {
-					response.sendRedirect("/WEB-INF/views/user/myinfo-m.jsp");
+					response.sendRedirect("/user/list?dogCheck=all");
 				}else {
 					System.out.println("관리자 메뉴에서 회원 삭제 오류 발생 (DogCheck == Y)");
 				}
